requirements.txt
# Scalable Food Tracker - Python Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Rate Limiting
fastapi-limiter==0.1.5

# Data Validation
pydantic==2.5.0

# Google Sheets API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Claude AI
anthropic==0.7.8

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Async Support
aiofiles==23.2.1
